+++ agent/bindings/kotlin/lib/src/main/kotlin/com/sourcegraph/cody/protocol_generated/WorkspaceEditMetadata.kt
@@ -1,7 +1,0 @@
-@file:Suppress("FunctionName", "ClassName", "unused", "EnumEntryName", "UnusedImport")
-package com.sourcegraph.cody.protocol_generated
-
-data class WorkspaceEditMetadata(
-  val isRefactoring: Boolean? = null,
-)
-