+++ vscode/src/non-stop/FixupController.ts
@@ -22,6 +22,7 @@ import { lines } from '../completions/text-processing'
 import { getInput } from '../edit/input/get-input'
 import type { ExtensionClient } from '../extension-client'
 import type { AuthProvider } from '../services/AuthProvider'
+import { isInTutorial } from '../tutorial/helpers'
 import { FixupDecorator } from './FixupDecorator'
 import { FixupDocumentEditObserver } from './FixupDocumentEditObserver'
 import type { FixupFile } from './FixupFile'
@@ -846,6 +847,13 @@ export class FixupController
         task: FixupTask,
         options?: { undoStopBefore: boolean; undoStopAfter: boolean }
     ): Promise<boolean> {
+        if (isInTutorial(document)) {
+            // Skip formatting in tutorial files,
+            // This is an additional enhancement that doesn't add much value to the tutorial
+            // and makes the tutorial UX more error-prone
+            return false
+        }
+
         // Expand the range to include full lines to reduce the likelihood of formatting issues
         const rangeToFormat = new vscode.Range(
             task.selectionRange.start.line,