+++ vscode/webviews/promptEditor/nodes/ContextItemMentionNode.ts
@@ -29,7 +29,7 @@ export const MENTION_CLASS_NAME = styles.contextItemMentionNode
  * The subset of {@link ContextItem} fields that we need to store to identify and display context
  * item mentions.
  */
-export type SerializedContextItem = { uri: string; content?: undefined } & (
+export type SerializedContextItem = { uri: string; title?: string; content?: undefined } & (
     | Omit<ContextItemFile, 'uri' | 'content'>
     | Omit<ContextItemSymbol, 'uri' | 'content'>
     | Omit<ContextItemPackage, 'uri' | 'content'>
@@ -179,8 +179,11 @@ export function contextItemMentionNodeDisplayText(contextItem: SerializedContext
     // A displayed range of `foo.txt:2-4` means "include all of lines 2, 3, and 4", which means the
     // range needs to go to the start (0th character) of line 5. Also, `RangeData` is 0-indexed but
     // display ranges are 1-indexed.
-    const rangeText = contextItem.range ? `:${displayLineRange(contextItem.range)}` : ''
+    const rangeText = contextItem.range?.start ? `:${displayLineRange(contextItem.range)}` : ''
     if (contextItem.type === 'file') {
+        if (contextItem.provider && contextItem.title) {
+            return `@${contextItem.title}`
+        }
         return `@${decodeURIComponent(displayPath(URI.parse(contextItem.uri)))}${rangeText}`
     }
     if (contextItem.type === 'symbol') {